requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==1.24.0
openai-whisper==20240930
packaging==25.0
pydub==0.25.1
regex==2024.11.6
requests==2.32.3
sympy==1.13.3
tiktoken==0.9.0
torch==2.2.2
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
